feat: 라우팅 구조 수정 및 공용 레이아웃 적용 범위 수정 (#20)
