--- src/app/(site)/layout.tsx
@@ -1,14 +1,0 @@
-import SiteHeader from '@/components/layout/header'
-
-export default function SiteLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <div className="min-h-screen">
-      <SiteHeader />
-      <main>{children}</main>
-    </div>
-  )
-}
