+++ src/app/components/reward-info.tsx
