+++ src/app/components/survey-card.tsx
@@ -3,11 +3,11 @@
 import Image from 'next/image'
 import Link from 'next/link'
 import { Card } from '@/components/ui/card'
-import { TagList } from '@/app/(site)/components/tag-list'
-import { CountdownTimer } from '@/app/(site)/components/countdown-timer'
-import { ProgressSection } from '@/app/(site)/components/progress-section'
-import { QuestionList } from '@/app/(site)/components/question-list'
-import { AnswerSection } from '@/app/(site)/components/answer-section'
+import { TagList } from '@/app/components/tag-list'
+import { CountdownTimer } from '@/app/components/countdown-timer'
+import { ProgressSection } from '@/app/components/progress-section'
+import { QuestionList } from '@/app/components/question-list'
+import { AnswerSection } from '@/app/components/answer-section'
 import { PointBadge, MinuteBadge, GifticonBadge } from '@/components/ui/badges'
 import { cn } from '@/lib/utils'
