+++ src/app/components/survey-preview-section.tsx
@@ -1,9 +1,9 @@
-import { SurveyCard } from '@/app/(site)/components/survey-card'
-import { SectionHeader } from '@/app/(site)/components/section-header'
+import { SurveyCard } from '@/app/components/survey-card'
+import { SectionHeader } from '@/app/components/section-header'
 import { ResponsiveSliderGrid } from '@/components/ui/responsive-slider-grid'
 import { SurveyItem } from '@/types/survey'
 
-export function NewestSurveySection({
+export function SurveyPreviewSection({
   items,
 }: {
   items: (SurveyItem & { answers: string[] })[]
@@ -11,8 +11,8 @@ export function NewestSurveySection({
   return (
     <section>
       <SectionHeader
-        title="방금 올라온 따끈따끈한 설문조사"
-        moreHref="/all-survey?tab=newest"
+        title="이런 설문조사는 어때요?"
+        moreHref="/all-survey"
       />
       <ResponsiveSliderGrid>
         {items.map((s) => (
@@ -23,9 +23,9 @@ export function NewestSurveySection({
             thumbnail={s.thumbnail}
             reward={s.reward}
             duration={s.duration}
-            tags={s.tags}
             variant="withAnswers"
             answers={s.answers}
+            tags={s.tags}
           />
         ))}
       </ResponsiveSliderGrid>
