+++ src/app/my-surveys/register/components/payment-button-section.tsx
