+++ src/app/my-surveys/register/components/platform-fee-section.tsx
