+++ src/app/my-surveys/register/components/point-payment-summary.tsx
