+++ src/app/my-surveys/register/components/registration-preview.tsx
