+++ src/app/my-surveys/register/components/reward-detail-section.tsx
