+++ src/app/page.tsx
@@ -1,11 +1,11 @@
-import { AdSection, BannerItem } from '@/app/(site)/components/ad-section'
-import { StatusSection } from '@/app/(site)/components/status-section'
-import { CustomSurveySection } from '@/app/(site)/components/custom-survey-section'
-import { UrgentCarouselSection } from '@/app/(site)/components/urgent-carousel-section'
-import { SurveyPreviewSection } from '@/app/(site)/components/survey-preview-section'
-import { WaitingSurveySection } from '@/app/(site)/components/waiting-survey-section'
-import { NewestSurveySection } from '@/app/(site)/components/newest-survey-section'
-import { BestRewardProductsSection } from '@/app/(site)/components/best-reward-products-section'
+import { AdSection, BannerItem } from '@/app/components/ad-section'
+import { StatusSection } from '@/app/components/status-section'
+import { CustomSurveySection } from '@/app/components/custom-survey-section'
+import { UrgentCarouselSection } from '@/app/components/urgent-carousel-section'
+import { SurveyPreviewSection } from '@/app/components/survey-preview-section'
+import { WaitingSurveySection } from '@/app/components/waiting-survey-section'
+import { NewestSurveySection } from '@/app/components/newest-survey-section'
+import { BestRewardProductsSection } from '@/app/components/best-reward-products-section'
 import { SurveyItem } from '@/types/survey'
 import { ProductItem } from '@/types/product'
